Adding Grpc connexion with CRUD methods

+++ OptionPricing.Domain/Strike.cs
@@ -8,9 +8,9 @@ namespace OptionPricing.Domain
 {
     public class Strike
     {
-        public double Value { get; private set; }
+        public float Value { get; private set; }
 
-        public Strike(double strike)
+        public Strike(float strike)
         {
             if (strike <= 0) throw new Exception("Negative values not callable... \n");
             Value = strike;
